requirements.txt
Django==2.2.6
coverage==4.5.4
psycopg2-binary==2.8.3
djangorestframework==3.10.3
celery==4.2.1
redis==3.3.11
wdb==3.3.0 
yarg==0.1.9
factory-boy==2.12.0
django-extensions==2.2.5
django-elasticsearch-dsl==6.4.1
mock==3.0.5
python-decouple==3.1
